+++ src/main/java/org/apache/hop/datavault/hopgui/file/dimensional/HopDimensionalFileType.java
@@ -27,13 +27,16 @@ import org.apache.hop.core.exception.HopException;
 import org.apache.hop.core.file.IHasFilename;
 import org.apache.hop.core.gui.plugin.action.GuiAction;
 import org.apache.hop.core.gui.plugin.action.GuiActionType;
+import org.apache.hop.core.search.ISearchable;
 import org.apache.hop.core.variables.IVariables;
 import org.apache.hop.core.vfs.HopVfs;
 import org.apache.hop.core.xml.XmlHandler;
 import org.apache.hop.datavault.hopgui.file.ExplorerPerspectiveTabSupport;
+import org.apache.hop.datavault.hopgui.search.HopGuiDimensionalModelSearchable;
 import org.apache.hop.datavault.metadata.ModelXmlWriteSupport;
 import org.apache.hop.datavault.metadata.dimensional.DimensionalModel;
 import org.apache.hop.datavault.metadata.dimensional.DmModelLoadSupport;
+import org.apache.hop.datavault.resourcedefinition.ResourceDefinitionGroupResolver;
 import org.apache.hop.i18n.BaseMessages;
 import org.apache.hop.metadata.api.IHopMetadataProvider;
 import org.apache.hop.metadata.serializer.xml.XmlMetadataUtil;
@@ -99,7 +102,21 @@ public class HopDimensionalFileType extends HopFileTypeBase {
     caps.setProperty(IHopFileType.CAPABILITY_PASTE, "true");
     caps.setProperty(IHopFileType.CAPABILITY_CUT, "true");
     caps.setProperty(IHopFileType.CAPABILITY_DELETE, "true");
+    caps.setProperty(IHopFileType.CAPABILITY_SEARCH, "true");
     return caps;
+  }
+
+  @Override
+  public ISearchable createSearchable(
+      String filename,
+      String locationDescription,
+      IVariables variables,
+      IHopMetadataProvider metadataProvider)
+      throws HopException {
+    DimensionalModel model =
+        ResourceDefinitionGroupResolver.loadDimensionalModel(
+            filename, variables, metadataProvider);
+    return new HopGuiDimensionalModelSearchable(locationDescription, model);
   }
 
   @Override