+++ src/main/java/org/apache/hop/datavault/hopgui/file/sourcemodel/HopSourceModelFileType.java
@@ -26,12 +26,15 @@ import org.apache.hop.core.exception.HopException;
 import org.apache.hop.core.file.IHasFilename;
 import org.apache.hop.core.gui.plugin.action.GuiAction;
 import org.apache.hop.core.gui.plugin.action.GuiActionType;
+import org.apache.hop.core.search.ISearchable;
 import org.apache.hop.core.variables.IVariables;
 import org.apache.hop.core.vfs.HopVfs;
 import org.apache.hop.core.xml.XmlHandler;
 import org.apache.hop.datavault.hopgui.file.ExplorerPerspectiveTabSupport;
+import org.apache.hop.datavault.hopgui.search.HopGuiSourceModelSearchable;
 import org.apache.hop.datavault.metadata.ModelXmlWriteSupport;
 import org.apache.hop.datavault.metadata.sourcemodel.SourceModel;
+import org.apache.hop.datavault.metadata.sourcemodel.SourceModelLoadSupport;
 import org.apache.hop.i18n.BaseMessages;
 import org.apache.hop.metadata.api.IHopMetadataProvider;
 import org.apache.hop.metadata.serializer.xml.XmlMetadataUtil;
@@ -96,7 +99,19 @@ public class HopSourceModelFileType extends HopFileTypeBase {
     caps.setProperty(IHopFileType.CAPABILITY_PASTE, "true");
     caps.setProperty(IHopFileType.CAPABILITY_CUT, "true");
     caps.setProperty(IHopFileType.CAPABILITY_DELETE, "true");
+    caps.setProperty(IHopFileType.CAPABILITY_SEARCH, "true");
     return caps;
+  }
+
+  @Override
+  public ISearchable createSearchable(
+      String filename,
+      String locationDescription,
+      IVariables variables,
+      IHopMetadataProvider metadataProvider)
+      throws HopException {
+    SourceModel model = SourceModelLoadSupport.load(filename, variables, metadataProvider);
+    return new HopGuiSourceModelSearchable(locationDescription, model);
   }
 
   @Override